second to last modification, last will be .md

+++ srcs/requirements/wordpress/start_wordpress.sh
@@ -3,24 +3,26 @@ echo wordpress startup
 
 cd /var/www/html/
 
-curl -O https://raw.githubusercontent.com/wp-cli/builds/gh-pages/phar/wp-cli.phar
-chmod +x wp-cli.phar
-mv wp-cli.phar /usr/local/bin/wp
+if [ ! -f /var/www/html/.init ]; then
+	touch /var/www/html/.init
+	echo Wordpress Init
 
-wp core download --allow-root
-cp /wp-config.php ./
-wp core install \
-	--url=$DOMAIN_NAME/ \
-	--title=$WP_TITLE \
-	--admin_user=$WP_ADMIN_USR \
-	--admin_password=$WP_ADMIN_PWD \
-	--admin_email=$WP_ADMIN_EML \
-	--allow-root
+	curl -O https://raw.githubusercontent.com/wp-cli/builds/gh-pages/phar/wp-cli.phar
+	chmod +x wp-cli.phar
+	mv wp-cli.phar /usr/local/bin/wp
 
-wp user create $WP_USR $WP_EML --role=author --user_pass=$WP_PWD --allow-root
+	wp core download --allow-root
+	cp /wp-config.php ./
+	wp core install \
+		--url=$DOMAIN_NAME/ \
+		--title=$WP_TITLE \
+		--admin_user=$WP_ADMIN_USR \
+		--admin_password=$WP_ADMIN_PWD \
+		--admin_email=$WP_ADMIN_EML \
+		--allow-root
 
-wp plugin update --all --allow-root
+	wp user create $WP_USR $WP_EML --role=author --user_pass=$WP_PWD --allow-root
+fi
 
-echo "done asdaaaaaaaaaaaaaaaaa"
-
+echo "Wordpress done"
 php-fpm8.2 -F
